+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
-# [v1.7.0](Unreleased)
+# [v1.7.0] (2020-11-05)
 ### Notes
 - This release extends supports of the SDK to Oneview API Version 2200 and Image Streamer API Version 2000.